[color] Explicitly cast trunc result to int [pimath] Rename from pmath, create stub [misc] Initialized README.md

+++ color.h
@@ -8,9 +8,9 @@
 #include "vec3.h"
 
 extern inline void write_color(struct vec3* pixel_color) {
-    int r = trunc(255.999 * pixel_color->x);
-    int g = trunc(255.999 * pixel_color->y);
-    int b = trunc(255.999 * pixel_color->z);
+    int r = (int)trunc(255.999 * pixel_color->x);
+    int g = (int)trunc(255.999 * pixel_color->y);
+    int b = (int)trunc(255.999 * pixel_color->z);
     printf("%d %d %d\n", r, g, b);
 }
